Crate ethcore_db

OpenEthereum database access utilities.

Modules

cache_manager

Database cache manager

keys

Blockchain DB extras.

Structs

DBTransaction

Write transaction. Batches a sequence of put/delete operations for efficiency.

DatabaseWithMetrics

Database with enabled statistics

InMemoryWithMetrics

InMemory with disabled statistics

Enums

CacheUpdatePolicy

Modes for updating caches.

Constants

COL_ACCOUNT_BLOOMDeprecated

Column for the accounts existence bloom filter.

COL_BODIES

Column for Block bodies

COL_EXTRA

Column for Extras

COL_HEADERS

Column for Block headers

COL_NODE_INFO

Column for general information from the local node which can persist.

COL_STATE

Column for State

COL_TRACE

Column for Traces

NUM_COLUMNS

Number of columns in DB

Traits

Cache

A cache for arbitrary key-value pairs.

Key

Should be used to get database key associated with given value.

KeyValueDB

Ethcore definition of a KeyValueDB with embeeded metrics

Readable

Should be used to read values from database.

Writable

Should be used to write value into database.

Type Definitions

DBValue

Database value.